fix: merge goal transitioner test to satisfy pkg file-count gate
